5월 26일
Dictionary를 Json화 시킬수있는 함수 추가
소환 컨텐츠 추가중

+++ Controller/TagController.cs
@@ -54,7 +54,7 @@ public class TagController : MonoBehaviour
             {
                 case OBJ_TYPE.PLAYER:
                     {
-                        hpGauge.fillAmount  = PlayerController.Instance.CurHP / PlayerController.Instance.MaxHP;
+                        hpGauge.fillAmount  = (float)(PlayerController.Instance.CurHP / PlayerController.Instance.MaxHP);
                     }
                     break;
             }
